pyproject.toml
[tool.poetry]
name = "arma_server_tools"
version = "0.1.2"
description = "Tools to manage content for arma3Server"
authors = ["dent <dent308actual@gmail.com>"]
repository = "https://github.com/dent308/arma_server_tools"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pyyaml = "^5.4.1"
click-log = "^0.3.2"
pytest = "^6.2.2"
rich = "^9.13.0"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]
click = "^7.1.2"
click-log = "^0.3.2"

[tool.poetry.scripts]
arma_server = "arma_server_tools.arma_server:main"
steam_pull = "arma_server_tools.workshop:main"
preset_parser = "arma_server_tools.preset_parser:main"
yaml_setup = "arma_server_tools.yaml_tools:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

